Add 'make check' recipe for automating building and running of tests.

+++ tests/macro_utils.h
@@ -2,12 +2,12 @@
 #define MACRO_UTILS_H
 
 #define expect_times(mocked_func, constraint, n) \
-    for(int i = 0; i < n; i++) { \
+    for(size_t i = 0; i < n; i++) { \
         expect(mocked_func, constraint); \
     }
 
 #define expect_many(mocked_func, constraint, arr) \
-    for(int i = 0; i < sizeof(arr); i++) { \
+    for(size_t i = 0; i < sizeof(arr); i++) { \
         expect(mocked_func, constraint(arr[i])); \
     }
